Windowing and window events implemented
Windowing is implemented via a factory pattern where Create is called in the App class and platform specific details are implenteded in child classes. The event system uses a lamda function of the OnEvent method in on update for all of the glfwcallbacks. inside OnEvent is a dispatcher that will check if the event is one the one that is desired and call the lamda function for said event handling. I did this all on main. oops