Basic Layer implementation
Layers are created via the Layer Interface which will hold OnUpdate, and OnEvent, functions which will be implemented per layer. When a new Layer child is created it is pushed into a LayerStack that owned by the RealApp engine class. The Stack will keep game layers and other lower render order objects at the top while overlays will be in the back to facilitate proper rendering order. During the Run() function in RealApp, the LayerStacks layers call their OnUpdate function. Same goes for the OnEvent as well. Feature needed to ignore layer via a bool