Input Polling implemented
simple Input polling added via an Input singleton static interface class that will be expanded to implementation per platform. The singleton is created in the platform implamentation file and is a polymorphic variation of input. When Input:: gets called it will return the values that the instance holds. This will need an implamentation of Real Engine key codes as to not uses GLFW in the application.